+++ lib/td_dd/systems/system_search.ex
@@ -30,6 +30,7 @@ defmodule TdDd.Systems.SystemSearch do
         id: system.id,
         name: system.name,
         external_id: system.external_id,
+        df_content: system.df_content,
         structures_count: Map.get(structures_count, "doc_count")
       }
     end)