+++ priv/repo/migrations/20200224094712_add_df_content_to_system.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddDfContentToSystem do
+  use Ecto.Migration
+
+  def change do
+    alter table(:systems) do
+      add :df_content, :map
+    end
+  end
+end